Fixed major bug, started ValString pooling
Before we had the pool as a static in the base class. This was incidentally sharing a pool across all children. Now the children all have their own pools. Messier, but it works as intended now. There's still an issue where a lot of the poolable types don't seem to be actually returning to the pool